+++ packages/scenes/src/variables/adhoc/controller/AdHocFiltersVariableController.ts
@@ -13,6 +13,7 @@ export class AdHocFiltersVariableController implements AdHocFiltersController {
 
   public useState(): AdHocFiltersControllerState {
     const state = this.model.useState();
+
     return {
       filters: state.filters,
       originFilters: state.originFilters,
@@ -22,6 +23,8 @@ export class AdHocFiltersVariableController implements AdHocFiltersController {
       onAddCustomValue: state.onAddCustomValue,
       wip: state._wip,
       collapsible: state.collapsible,
+      valueRecommendations: this.model.getRecommendations(),
+      drilldownRecommendationsEnabled: state.drilldownRecommendationsEnabled,
     };
   }
 
@@ -39,6 +42,16 @@ export class AdHocFiltersVariableController implements AdHocFiltersController {
 
   public updateFilter(filter: AdHocFilterWithLabels, update: Partial<AdHocFilterWithLabels>): void {
     this.model._updateFilter(filter, update);
+  }
+
+  public updateFilters(
+    filters: AdHocFilterWithLabels[],
+    options?: {
+      skipPublish?: boolean;
+      forcePublish?: boolean;
+    }
+  ): void {
+    this.model.updateFilters(filters, options);
   }
 
   public updateToMatchAll(filter: AdHocFilterWithLabels): void {